Refactored notification functionality into its own package. Fixed issue where system notification was being sent each time an api call was made. Updated MainView to update the Gas price label with each query.

+++ src/main/kotlin/com/mpcrypto/ethereumgaschecker/fileio/Preferences.kt
@@ -5,5 +5,5 @@ package com.mpcrypto.ethereumgaschecker.fileio
  */
 data class Preferences(
     var durationThreshold : Int,
-    var gasThreshold : Double,
+    var gasThreshold : Int,
     var scanFrequency : Int)
